+++ packages/core/services/logService/logService.test.js
@@ -4,14 +4,13 @@ describe('logService', () => {
   it('should return an object of the correct shape', () => {
     const service = getService('irving:test');
     expect(Object.keys(service)).toMatchObject([
-      'emerg',
-      'alert',
-      'crit',
       'error',
-      'warning',
-      'notice',
+      'warn',
       'info',
+      'http',
+      'verbose',
       'debug',
+      'silly',
     ]);
   });
 });