+++ packages/core/services/monitorService/defaultService.js
@@ -1,5 +1,6 @@
 module.exports = () => ({
   start: () => {},
   logError: () => {},
+  logMessage: () => {},
   logTransaction: () => {},
 });