+++ packages/integrations/components/pico/useGadgetScript.js
@@ -2,6 +2,7 @@
 import { useEffect, useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import getLogService from '@irvingjs/services/logService';
+import getEnv from '@irvingjs/core/config/multisite/getEnv';
 import { actionUpdatePicoLifecycle } from '../../actions/picoActions';
 import { picoLifecycleSelector } from '../../selectors/picoSelector';
 
@@ -9,6 +10,10 @@ const log = getLogService('irving:pico');
 
 export default function useGadgetScript(gadgetUrl, publisherId) {
   const dispatch = useDispatch();
+
+  // Construct script source.
+  const { PICO_SCRIPT_FILENAME } = getEnv();
+  const scriptFilename = PICO_SCRIPT_FILENAME || 'wrapper.min.js';
 
   // Create a function that updates the store whenever the script is added.
   const dispatchUpdateScriptAdded = useCallback(
@@ -33,7 +38,7 @@ export default function useGadgetScript(gadgetUrl, publisherId) {
 
       // Set the attributes and content.
       script.type = 'text/javascript';
-      script.text = `(function(p,i,c,o){var n=new Event("pico.scriptOnload");i[p]=i[p]||function(){(i[p].queue=i[p].queue||[]).push(arguments)},i.document.addEventListener("pico.scriptOnload",function(e){var t=i.Pico.getInstance(e,{publisherId:o,picoInit:n},i);t.handleQueueItems(i[p].queue),i[p]=function(){return t.handleQueueItems([arguments])}},!1);var e=i.document.createElement("script"),t=i.document.getElementsByTagName("script")[0];e.async=1,e.defer=1,e.src=c,e.onload=function(e){return i.Pico.getInstance(e,{publisherId:o,picoInit:n},i)},t.parentNode.insertBefore(e,t)})("pico",window,"${gadgetUrl}/wrapper.min.js", "${publisherId}");`;
+      script.text = `(function(p,i,c,o){var n=new Event("pico.scriptOnload");i[p]=i[p]||function(){(i[p].queue=i[p].queue||[]).push(arguments)},i.document.addEventListener("pico.scriptOnload",function(e){var t=i.Pico.getInstance(e,{publisherId:o,picoInit:n},i);t.handleQueueItems(i[p].queue),i[p]=function(){return t.handleQueueItems([arguments])}},!1);var e=i.document.createElement("script"),t=i.document.getElementsByTagName("script")[0];e.async=1,e.defer=1,e.src=c,e.onload=function(e){return i.Pico.getInstance(e,{publisherId:o,picoInit:n},i)},t.parentNode.insertBefore(e,t)})("pico",window,"${gadgetUrl}/${scriptFilename}", "${publisherId}");`;
 
       // Append the script into the root of the body.
       documentBody.appendChild(script);