+++ packages/vip-go/config/webpack.config.client.js
@@ -1,7 +1,7 @@
 module.exports = {
   resolve: {
     fallback: {
-      buffer: require.resolve('buffer'),
+      util: require.resolve('util'),
     },
   },
 };