+++ packages/core/CHANGELOG.md
@@ -3,6 +3,74 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.2.0-alpha.0](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.4...v6.2.0-alpha.0) (2021-05-11)
+
+
+
+# [6.1.0-alpha.7](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-alpha.6...v6.1.0-alpha.7) (2021-05-03)
+
+**Note:** Version bump only for package @irvingjs/core
+
+
+
+
+
+# [6.1.0-alpha.7](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-alpha.6...v6.1.0-alpha.7) (2021-05-03)
+
+**Note:** Version bump only for package @irvingjs/core
+# [6.1.0](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-beta.1...v6.1.0) (2021-05-04)
+
+**Note:** Version bump only for package @irvingjs/core
+
+
+
+
+
+# [6.1.0-beta.1](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-beta.0...v6.1.0-beta.1) (2021-05-04)
+
+
+### Bug Fixes
+
+* **logservice:** normalize log messages and errors ([38a6bcf](https://github.com/alleyinteractive/irving/packages/core/commit/38a6bcfa2b5466296235d6df974104e39edb7d6c))
+
+
+
+
+
+# [6.1.0-beta.0](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-alpha.6...v6.1.0-beta.0) (2021-05-03)
+
+**Note:** Version bump only for package @irvingjs/core
+
+
+
+
+
+# [6.1.0-alpha.6](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-alpha.5...v6.1.0-alpha.6) (2021-05-03)
+
+
+### Bug Fixes
+
+* **pico:** add new PICO_SCRIPT_FILENAME environmental variable ([c82fed0](https://github.com/alleyinteractive/irving/packages/core/commit/c82fed052cda01a365320f9d5e7cb7e46314a805))
+
+
+
+
+
+# [6.1.0-alpha.4](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-alpha.3...v6.1.0-alpha.4) (2021-04-28)
+
+
+### Bug Fixes
+
+* **logging:** make sure emerg, crit, and error log methods use Error object ([3d96885](https://github.com/alleyinteractive/irving/packages/core/commit/3d968854b2cf7ba56d0fd90d8940882c8a79e4cc))
+* **logservice:** add back in log for sending logs to newrelic monitoring ([5e2ef43](https://github.com/alleyinteractive/irving/packages/core/commit/5e2ef43eacd0728d8daa00701dff862277442e8f))
+* **logservice:** remove unnecessary newrelic transport, fix eslint issue ([0a14ce2](https://github.com/alleyinteractive/irving/packages/core/commit/0a14ce288cd29645ac16d7862f42aefa377057b5))
+* **logservice:** switch up log levels to match winston ([cbb88db](https://github.com/alleyinteractive/irving/packages/core/commit/cbb88db3cccc35d76826ff19d8a76ee11076eacd))
+* **test:** fix service tests ([63cd51e](https://github.com/alleyinteractive/irving/packages/core/commit/63cd51e2874e3f75f94a1b313b71970972343bb0))
+
+
+
+
+
 # [6.1.0-alpha.2](https://github.com/alleyinteractive/irving/packages/core/compare/v6.1.0-alpha.1...v6.1.0-alpha.2) (2021-04-23)